requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
torch==2.1.0
torchvision==0.16.0
timm==0.9.12

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Data Science
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
